Use new container for level strings
