+++ src/Engine/Events.cpp
@@ -5,16 +5,6 @@
 
 // TODO(Nik-RE-dev): move remaining globals
 
-std::array<unsigned int, 500> pLevelStrOffsets;
-unsigned int uLevelStrNumStrings;
-unsigned int uLevelStrFileSize;
-std::array<char, 9216> pLevelStr;
-
 _2devent p2DEvents[525];
 
 MapEventVariables mapEventVariables;
-
-int savedEventID;
-int savedEventStep;
-struct LevelDecoration *savedDecoration;
-
